fixed issue with closing files on exit

+++ src/userprog/syscall.c
@@ -282,9 +282,9 @@ int read(const uint8_t* stack){
     lock_release(&file_lock);
     return -1;
   }
-  size = (int)file_read(f->file, buffer, (off_t)size);
+  size = file_read(f->file, buffer, (off_t)size);
   lock_release(&file_lock);
-  return size;
+  return (int)size;
 }
 
 /*Handler for SYS_WRITE*/
